Initial version of backend part:
Add:
1. Model:
ImportConfiguration - contains info required to import the entities for specified metaclass.
PropertyMapping - contains info on how to map the field from source (XLSX, JSON, XML, CSV) and entity property.
PropertyMapping can be set for simple property or for reference entity.
The association for the reference entity is processed according to the specified ReferenceEntityPolicy.

UniqueEntityConfiguration - contains info by which properties the entities to import are considered as identical.
If two entities have the same values of specified properties,
the duplicate entity will be processed according to the specified policy.

2. Beans:
 a) DataImporter - the main API to import data from the specified source (string - for XML, JSON, CSV only, input stream, or imported data) using import configuration.
 b) DataExtractor - extracts the data from source (XML, JSON, XLSX or CSV).
 c) EntityExtractor - creates an entity to import and set values of the properties for which mapping is set.
 d) ImportConfigurationBuilder, ImportConfigurationBuilders - helper classes to build ImportConfiguration.
 e) DuplicateEntityManager - to search the duplicate entities
 f) PropertyBinder - implementations of this interface are used to get a property value and set it for the specified entity.
 g) AssociationCreator - creates an association for the property from mapping.